🏗️ Split `feed` selectors into specific feature folders
In the new feature-based architecture, each feature has its own dedicated `store` subfolder. Selectors are the first part migrated into it.
