+++ src/features/activity-feed/comment/store/commentsSelectors.js
@@ -1,0 +1,20 @@
+import { createSelector } from '@reduxjs/toolkit'
+import { selectPost } from '../../post/store/postsSelectors'
+
+/**
+ * @param {Object} state
+ * @param {string} postId
+ */
+export const selectPostCommentsCount = createSelector(
+    [selectPost],
+    post => post.commentsCount
+)
+
+/**
+ * @param {Object} state
+ * @param {string} postId
+ */
+export const selectPostComments = createSelector(
+    [selectPost],
+    post => post.comments
+)
