+++ src/features/activity-feed/post/components/ActivityFeed/index.jsx
@@ -2,7 +2,7 @@ import { useContext, useEffect } from 'react'
 import { useSelector, useDispatch } from 'react-redux'
 import { Box, Grid, CircularProgress } from '@mui/material'
 import { PostIdProvider } from '../../contexts/PostIdProvider'
-import { selectPosts, selectActivityFeedHasMorePosts, selectPostLoading } from '../../../../../redux/selectors/feed'
+import { selectPosts, selectActivityFeedHasMorePosts, selectPostLoading } from '../../store/postsSelectors'
 import { fetchPosts } from '../../../../../redux/thunks/feed'
 import { cleanFeedState } from '../../../../../redux/reducers/feed'
 import { ProfileUserIdContext } from '../../contexts/ProfileUserIdProvider'
