+++ src/features/activity-feed/post/components/Post/index.jsx
@@ -5,7 +5,8 @@ import moment from 'moment'
 import { Card, CardActions, CardHeader, CardContent, Typography, Button, Divider, Avatar, Collapse, Link as MuiLink } from '@mui/material'
 import { HashLink } from 'react-router-hash-link'
 import { PostIdContext } from '../../contexts/PostIdProvider'
-import { selectPost, selectPostReactions } from '../../../../../redux/selectors/feed'
+import { selectPost } from '../../store/postsSelectors'
+import { selectPostReactions } from '../../../reaction/store/reactionsSelectors'
 import { selectUser } from '../../../../user/store/userSelectors'
 import { CommentsList } from '../../../comment/components/CommentsList'
 import { CommentForm } from '../../../comment/components/CommentForm'
