+++ src/features/activity-feed/reaction/components/ReactionButton/index.jsx
@@ -4,7 +4,7 @@ import Typography from '@mui/material/Typography'
 import Button from '@mui/material/Button'
 import { useDispatch, useSelector } from 'react-redux'
 import { PostIdContext } from '../../../post/contexts/PostIdProvider'
-import { selectPostReactions } from '../../../../../redux/selectors/feed'
+import { selectPostReactions } from '../../store/reactionsSelectors'
 import { selectUserId } from '../../../../user/store/userSelectors'
 import { createReaction, updateReaction, removeReaction } from '../../../../../redux/thunks/feed'
 import './style.scss'
